AWS-Region option added
Added to config (.ini) parameter AWS_REGION to be able to choose a
different AWS region.
Necessary code changes to use this parameter are included as well.

the sqs_connection() call had to be changed to a
boto.sqs.connecto_to_region() call in case a specific AWS region is defined.